Switching the default slack in signature requests to 100MB.
This is per request of @synctext. Allowing up to
100MB to be unsigned will allow the multichain to grow
at a more reasonable pace.

In addition, it also fixes Test.test_multi_chain_community
test_schedule_block and test_schedule_block_negative,
as these tests seem to have been written with a 5MB cap
in mind, instead of the 1MB that is currently in the code.